Use "createTime" instead of "createDate"
I was getting "Invalid date" as creation date in the interface, because
the Javascript didn't seem to like the format of the dates returned by
BBB.

Switch to using the "createTime" attribute which is a standard Unix
timestamp.  It avoids the need to parse dates in Javascript.

Tested with BBB 2.2